APM-2975 Organise tests
