+++ envs.sh
@@ -1,0 +1,17 @@
+export APIGEE_ENVIRONMENT="internal-dev"
+export OAUTH_BASE_URI="https://internal-dev.api.service.nhs.uk"
+export OAUTH_PROXY="oauth2-mock"
+export PROXY_NAME="identity-service-mock-pr-308"
+export API_NAME="identity-service"
+export SERVICE_NAME="identity-service"
+export ID_TOKEN_PRIVATE_KEY_ABSOLUTE_PATH="id-token-testing.key"
+export ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH="id-token-key.key"
+export JWT_PRIVATE_KEY_ABSOLUTE_PATH="jwt-testing-key.key"
+export APIGEE_ACCESS_TOKEN="$(get_token)"
+export APIGEE_API_TOKEN="$(get_token)"
+export STATUS_ENDPOINT_API_KEY="3fed41a7-1de3-4e03-980b-5945a4c06c86"
+export FULLY_QUALIFIED_SERVICE_NAME="identity-service-mock-pr-308"
+# export AUTHENTICATE_URL="https://internal-dev.api.service.nhs.uk/oauth2-mock"
+# export CLIENT_ID="test-client-1"
+# export CLIENT_SECRET="a9f1a286-2423-4451-99f7-c59a628c79ad"
+# export REDIRECT_URI="12345"
